Fix `failed to unpack tree object`; remove gopkg
Removes gopkg.in from import paths and re-initializes Gopkg.{toml,lock}.
This fixes an apparent problem with git.

While building I ran over this git error:

> ✔ /go/configmap-reload/src/github.com/jimmidyson/configmap-reload [master|✔]
> $ git describe
> v0.2.2-1-g633f4c9
> ✔ /go/configmap-reload/src/github.com/jimmidyson/configmap-reload [master|✔]
> $ git clean -x -f -d
> ✔ /go/configmap-reload/src/github.com/jimmidyson/configmap-reload [master|✔]
> $ dep ensure --vendor-only
> grouped write of manifest, lock and vendor: error while writing out vendor tree: failed to write dep tree: failed to export gopkg.in/fsnotify.v1: fatal: failed to unpack tree object c2828203cd70a50dcccfb2761f8b1f8ceef9a8e9
> : exit status 128
> $ dep version
> dep:
>  version     : devel
>  build date  :
>  git hash    :
>  go version  : devel +1a27f048ad Fri Jun 29 15:20:52 2018 +0000
>  go compiler : gc
>  platform    : linux/amd64
>  features    : ImportDuringSolve=false

Anyway, removing gopkg.in from import paths and doing a `dep init` from
scratch fixed it for me, so I did not bother to get to the bottom of
this.

IMHO gopkg.in is obsolete when versioning is managed by a tool like dep.

(I've manually reverted a change to the locked commit of golang/x/sys
that was introduced by `dep init`. So this commit is only about
gopkg.in; other dependencies should not be affected.)